+++ src/sites/image/reklama.js
@@ -154,7 +154,7 @@
       host: /^imgprime\.com$/,
       path: /^\/imga-u\/(.+)\.jpeg\.html/,
     },
-    async ready () {
+    async start () {
       const path = window.location.href.replace('/imga-u', '/u').replace('.html', '');
       await $.openLink(path);
     },
@@ -163,10 +163,10 @@
   _.register({
     rule: {
       host: /^22pixx\.xyz$/,
-      path: /^\/ia-i\/(.+)\.jpeg\.html/,
+      path: /^\/ia-[io]\/(.+)\.jpeg\.html/,
     },
-    async ready () {
-      const path = window.location.href.replace('/ia-i', '/i').replace('.html', '');
+    async start () {
+      const path = window.location.href.replace('/ia-', '/').replace('.html', '');
       await $.openLink(path);
     },
   });
@@ -176,7 +176,7 @@
       host: /^22pixx\.xyz$/,
       path: /^\/x-i\/(.+)\.jpeg\.html/,
     },
-    async ready () {
+    async start () {
       const path = window.location.href.replace('/x', '/y');
       await $.openLink(path);
     },
@@ -244,24 +244,26 @@
   async function action (firstSelector, secondSelector) {
     $.remove('iframe, #adblock_detect, .popupOverlay');
 
-    const node = $.$(firstSelector);
+    let node = $.$(firstSelector);
     if (node) {
-      // first pass
-      await _.wait(500);
-      node.removeAttribute('disabled');
-      await _.wait(500);
-      // HACK some sites can not receive the click event without focus
-      node.focus();
-      // HACK some sites needs to click multiple times
-      node.click();
-      node.click();
+      node = findFirstForm(node);
+      // clone the form and replace it to body to strip events
+      document.body.innerHTML = node.outerHTML;
+      node = $('form input');
       node.click();
       return;
     }
 
     // second pass
-    const i = $(secondSelector);
-    await $.openImage(i.src);
+    node = $(secondSelector);
+    await $.openImage(node.src);
+  }
+
+  function findFirstForm (child) {
+    while (child && child.localName !== 'form') {
+      child = child.parentElement;
+    }
+    return child;
   }
 
 })();